+++ app/api/webforms/partnershipform/route.ts
@@ -35,6 +35,7 @@ export async function POST(req: Request) {
         contactName: formData.contactName,
         phoneNumber: formData.phoneNumber,
         emailAddress: formData.emailAddress,
+        status: "Delivered",
       },
     });
     return NextResponse.json({ status: 200 });
@@ -62,20 +63,41 @@ export async function GET() {
 }
 
 //develop a method that will delete the specified partnership form
-export async function DELETE(req: Request) {
-  try {
-    const formData = await req.json();
-    await prisma.partnershipForm.delete({
-      where: {
-        id: formData.id,
-      },
-    });
-    return NextResponse.json({ status: 200 });
-  } catch (e) {
-    console.error("ERROR: ", e);
-    return NextResponse.json(
-      { error: "Internal Server Error" },
-      { status: 500 }
-    );
-  }
+export async function PATCH(req: Request) {
+    try {
+        const formData = await req.json();
+        if (formData.event === 'delete') {
+            await prisma.partnershipForm.delete({
+                where: {
+                    id: formData.id
+                }
+            });
+        }
+        else if (formData.event === 'edit') {
+            await prisma.partnershipForm.update({
+                where: {
+                  id: formData.id
+                },
+                data: {
+                  status: formData.status
+                }
+            });
+        }
+        else if (formData.event === 'single'){
+            const form = await prisma.partnershipForm.findUnique({
+                where: {
+                    id: formData.id
+                }
+            });
+            return NextResponse.json(form);
+        }
+        else {
+            throw new Error("Invalid event type.");
+        }
+        return NextResponse.json({ status: 200 });
+    }
+    catch(e) {
+        console.error("ERROR: ", e);
+        return NextResponse.json({ error: 'Internal Server Error' }, { status: 500 });
+    }
 }
